+++ src/geometry_rapidrotator.jl
@@ -77,7 +77,7 @@ end
 # end
 
 # von Zeipel law
-@views function temperature_map_vonZeipel_rapid_rotator(stellar_parameters,star; offsets = [0.0,0.0,0.0], GM = 1.0, T=Float32)
+@views function temperature_map_vonZeipel_rapid_rotator(stellar_parameters,star; offsets = [0.0,0.0,0.0], GM = 1.0, T=Float64)
   toff= T.(offsets)'
   GM = T(GM)
   r_theta = sqrt.(dropdims(sum(abs2, (star.vertices_xyz[:,5,:] .- toff), dims=2), dims=2));
